Move command tree to separate file
